refactor: slpit installer.js
