+++ lib/index.js
@@ -1,10 +1,7 @@
 const core = require("@actions/core");
-const tc = require("@actions/tool-cache");
-const io = require("@actions/io");
 const exec = require("@actions/exec");
-const getOS = require("./get-os");
-const getURL = require("./get-url");
 const getLatestVersion = require("./get-latest-version");
+const installer = require("./installer");
 
 // most @actions toolkit packages have async methods
 async function run() {
@@ -17,30 +14,7 @@ async function run() {
         }
         console.log(`Hugo version: ${hugoVersion}`);
 
-        const extended = core.getInput("extended");
-        console.log(`Hugo extended: ${extended}`);
-
-        const osName = getOS(process.platform);
-        console.log(`Operating System: ${osName}`);
-
-        const hugoURL = getURL(osName, extended, hugoVersion);
-        core.debug(`hugoURL: ${hugoURL}`);
-
-        const hugoPath = `${process.env.HOME}/bin`;
-        await io.mkdirP(hugoPath);
-        core.addPath(hugoPath);
-
-        // Download and extract Hugo binary
-        const hugoAssets = await tc.downloadTool(hugoURL);
-        let hugoBin = "";
-        if (osName === "Windows") {
-          const hugoExtractedFolder = await tc.extractZip(hugoAssets, "/tmp");
-          hugoBin = `${hugoExtractedFolder}/hugo.exe`;
-        } else {
-          const hugoExtractedFolder = await tc.extractTar(hugoAssets, "/tmp");
-          hugoBin = `${hugoExtractedFolder}/hugo`;
-        }
-        await io.mv(hugoBin, hugoPath);
+        await installer(hugoVersion);
 
         // Show version
         await exec.exec("hugo version");
